+++ src/main/java/com/StudyMate/StudyMate/model/User.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Id;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String subscription;
     private String name;
     private String email;
@@ -17,18 +17,18 @@ public class User {
     public User() {
     }
 
-    public User(Integer id, String subscription, String name, String email) {
+    public User(Long id, String subscription, String name, String email) {
         this.id = id;
         this.subscription = subscription;
         this.name = name;
         this.email = email;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
